implement 'rm'
Unlike earlier iteration of spec this doesn't flag blob's for removal. Instead,
it unlinks the named files but not the SUM file unless it's one of the
requested files and the garbage-collector removes all singularly linked SUM
files.

Also, write potentially large output exec commands to a temp ACK file which
eliminates a copy.

TODO: add a `restore` command to recreate links to SUM files.

Signed-off-by: Tom Grennan <tom@apptimist.co>